Gộp code của Công, Done 50% (2)

+++ ASP_Hotel/App_Code/Class/bills.cs
@@ -17,7 +17,7 @@ public class bills
 	public int bill_id { get; set; }
 	public int schedule_id { get; set; }
 	public int num_day { get; set; }
-	public float price_room { get; set; }
+	public double price_room { get; set; }
 	public double price_service { get; set; }
 	public double total_price { get; set; }
 	public Boolean bill_status { get; set; }
